backend/requirements.txt
# ================================
# スタートアップウェルネス分析システム
# アプリケーション依存ライブラリ
# ================================
# ----- Webフレームワーク -----
fastapi>=0.112.0
uvicorn>=0.15.0
pydantic>=2.5.0
starlette>=0.17.1
dash==2.9.3
gunicorn==21.2.0
# ----- データベース -----
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
alembic>=1.7.3
# ----- データ処理拡張 -----
# 注: 基本ライブラリ(numpy/pandas/scipy)はenvironment.ymlで管理
openpyxl==3.1.2
db-dtypes==1.1.1
# ----- 統計・機械学習拡張 -----
# 注: scikit-learnはenvironment.ymlで管理
statsmodels==0.14.0
# ----- 因果推論・高度分析 -----
arviz==0.16.1
causalimpact==0.2.6
dowhy
econml==0.14.1
pgmpy==0.1.22
cvxpy==1.3.1
# ----- 時系列分析 -----
prophet==1.1.4
# ----- テキスト分析・NLP -----
nltk==3.8.1
textblob==0.17.1
# ----- 特殊分析モジュール -----
lifelines==0.27.4  # 生存時間分析
mlxtend==0.22.0    # アソシエーション分析
# ----- ベイズ推論拡張 -----
# 注: pyro-pplはenvironment.ymlで管理
pymc==5.6.1
# ----- グラフ解析 -----
# 注: networkxはenvironment.ymlで管理
neo4j
# ----- 可視化 -----
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
# ----- クラウド連携 -----
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
google-cloud-scheduler==2.11.0
google-cloud-monitoring==2.14.2
google-cloud-secret-manager==2.16.3
# ----- レポート生成 -----
reportlab==4.0.4
jinja2>=3.0.1
weasyprint==59.0
# ----- Google API連携 -----
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
# ----- 生成AI連携 -----
google-generativeai>=0.4.0
# ----- Webスクレイピング -----
requests>=2.26.0
beautifulsoup4==4.12.2
httpx==0.24.1
# ----- セキュリティ -----
pycryptodome==3.18.0
bcrypt>=3.2.0
authlib==1.2.1
argon2-cffi==23.1.0
python-jose>=3.3.0
python-multipart>=0.0.5
email-validator==1.3.1
pyotp==2.8.0
qrcode==7.4.2
cryptography==41.0.3
# ----- 連合学習拡張 -----
# 注: 基本パッケージ(flwr)はenvironment.ymlで管理
opacus==1.4.0  # 差分プライバシー
# ----- モニタリング・評価 -----
prometheus-client==0.17.1
optuna==3.3.0
great_expectations==0.17.19
# ----- ロギング・診断 -----
loguru>=0.5.3
pykwalify==1.8.0
# ----- ツール・ユーティリティ -----
boto3==1.28.19
celery==5.4.0
datashader==0.15.2
# ----- 開発・テスト -----
pytest>=6.2.5
pytest-asyncio>=0.15.1
flake8==6.1.0
black==23.7.0
mypy==1.5.1
# ----- フレームワーク連携 -----
apispec>=6.0.0,<7  # flask-appbuilderの要件に適合
flask-appbuilder==4.3.5  # apache-airflowの依存関係
# ----- 注意事項 -----
# - このファイルはアプリケーション固有のライブラリに特化しています
# - システム依存パッケージは environment.yml で管理してください
# - バージョン競合に注意:
#   * tensorflow-federated と apache-airflow は attrs のバージョンが競合
#   * apache-airflowは個別にインストール: AIRFLOW_GPL_UNIDECODE=yes pip install apache-airflow==2.7.1
# 追加の依存関係
python-magic>=0.4.24
aiofiles>=0.7.0
numpy>=1.21.2
pandas>=1.3.3
scikit-learn>=0.24.2
tensorflow>=2.9.0
tensorflow-probability>=0.17.0
keras>=2.9.0
tf-keras>=2.9.0
flwr>=1.0.0
pillow>=9.0.0
asyncio>=3.4.3
aiohttp>=3.8.1
python-dotenv>=0.19.0
stripe>=2.60.0
passlib>=1.7.4